# Testing: Speed up Docker setup on Travis for e2e tests