--- a/.travis.yml
+++ b/.travis.yml
@@ -22,37 +22,32 @@ branches:
   only:
     - master
 
+before_install:
+  - nvm install --latest-npm
+
 env: PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
 
 jobs:
   include:
     - name: Lint
-      before_install:
-        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run lint
 
     - name: Build artifacts
-      before_install:
-        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run check-local-changes
 
     - name: License compatibility
-      before_install:
-        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run check-licenses
 
     - name: JavaScript unit tests
-      before_install:
-        - nvm install --latest-npm
       install:
         - npm ci
       script:
@@ -88,97 +83,65 @@ jobs:
 
     - name: E2E tests (Admin with plugins) (1/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
-      before_install:
-        - nvm install --latest-npm
       install:
-        - npm ci
         - ./bin/setup-travis-e2e-tests.sh
       script:
-        - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
         - npm run test-e2e -- --ci --cacheDirectory="$HOME/.jest-cache" --runTestsByPath $( awk 'NR % 4 == 0' < ~/.jest-e2e-tests )
 
     - name: E2E tests (Admin with plugins) (2/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
-      before_install:
-        - nvm install --latest-npm
       install:
-        - npm ci
         - ./bin/setup-travis-e2e-tests.sh
       script:
-        - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
         - npm run test-e2e -- --ci --cacheDirectory="$HOME/.jest-cache" --runTestsByPath $( awk 'NR % 4 == 1' < ~/.jest-e2e-tests )
 
     - name: E2E tests (Admin with plugins) (3/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
-      before_install:
-        - nvm install --latest-npm
       install:
-        - npm ci
         - ./bin/setup-travis-e2e-tests.sh
       script:
-        - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
         - npm run test-e2e -- --ci --cacheDirectory="$HOME/.jest-cache" --runTestsByPath $( awk 'NR % 4 == 2' < ~/.jest-e2e-tests )
 
     - name: E2E tests (Admin with plugins) (4/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
-      before_install:
-        - nvm install --latest-npm
       install:
-        - npm ci
         - ./bin/setup-travis-e2e-tests.sh
       script:
-        - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
         - npm run test-e2e -- --ci --cacheDirectory="$HOME/.jest-cache" --runTestsByPath $( awk 'NR % 4 == 3' < ~/.jest-e2e-tests )
 
     - name: E2E tests (Author without plugins) (1/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
-      before_install:
-        - nvm install --latest-npm
       install:
-        - npm ci
         - ./bin/setup-travis-e2e-tests.sh
       script:
-        - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
         - npm run test-e2e -- --ci --cacheDirectory="$HOME/.jest-cache" --runTestsByPath $( awk 'NR % 4 == 0' < ~/.jest-e2e-tests )
 
     - name: E2E tests (Author without plugins) (2/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
-      before_install:
-        - nvm install --latest-npm
       install:
-        - npm ci
         - ./bin/setup-travis-e2e-tests.sh
       script:
-        - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
         - npm run test-e2e -- --ci --cacheDirectory="$HOME/.jest-cache" --runTestsByPath $( awk 'NR % 4 == 1' < ~/.jest-e2e-tests )
 
     - name: E2E tests (Author without plugins) (3/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
-      before_install:
-        - nvm install --latest-npm
       install:
-        - npm ci
         - ./bin/setup-travis-e2e-tests.sh
       script:
-        - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
         - npm run test-e2e -- --ci --cacheDirectory="$HOME/.jest-cache" --runTestsByPath $( awk 'NR % 4 == 2' < ~/.jest-e2e-tests )
 
     - name: E2E tests (Author without plugins) (4/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
-      before_install:
-        - nvm install --latest-npm
       install:
-        - npm ci
         - ./bin/setup-travis-e2e-tests.sh
       script:
-        - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
         - npm run test-e2e -- --ci --cacheDirectory="$HOME/.jest-cache" --runTestsByPath $( awk 'NR % 4 == 3' < ~/.jest-e2e-tests )
 
@@ -193,4 +156,3 @@ jobs:
       env: WP_VERSION=latest SWITCH_TO_PHP=5.2
       script:
         - ./bin/run-wp-unit-tests.sh
-