--- a/.travis.yml
+++ b/.travis.yml
@@ -22,32 +22,37 @@ branches:
   only:
     - master
 
-before_install:
-  - nvm install --latest-npm
-
 env: PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
 
 jobs:
   include:
     - name: Lint
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run lint
 
     - name: Build artifacts
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run check-local-changes
 
     - name: License compatibility
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run check-licenses
 
     - name: JavaScript unit tests
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
       script:
@@ -83,6 +88,8 @@ jobs:
 
     - name: E2E tests (Admin with plugins) (1/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
         - ./bin/setup-travis-e2e-tests.sh
@@ -93,6 +100,8 @@ jobs:
 
     - name: E2E tests (Admin with plugins) (2/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
         - ./bin/setup-travis-e2e-tests.sh
@@ -103,6 +112,8 @@ jobs:
 
     - name: E2E tests (Admin with plugins) (3/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
         - ./bin/setup-travis-e2e-tests.sh
@@ -113,6 +124,8 @@ jobs:
 
     - name: E2E tests (Admin with plugins) (4/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
         - ./bin/setup-travis-e2e-tests.sh
@@ -123,6 +136,8 @@ jobs:
 
     - name: E2E tests (Author without plugins) (1/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
         - ./bin/setup-travis-e2e-tests.sh
@@ -133,6 +148,8 @@ jobs:
 
     - name: E2E tests (Author without plugins) (2/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
         - ./bin/setup-travis-e2e-tests.sh
@@ -143,6 +160,8 @@ jobs:
 
     - name: E2E tests (Author without plugins) (3/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
         - ./bin/setup-travis-e2e-tests.sh
@@ -153,6 +172,8 @@ jobs:
 
     - name: E2E tests (Author without plugins) (4/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
+      before_install:
+        - nvm install --latest-npm
       install:
         - npm ci
         - ./bin/setup-travis-e2e-tests.sh