--- a/.travis.yml
+++ b/.travis.yml
@@ -23,39 +23,31 @@ branches:
     - master
 
 before_install:
-  - nvm install
+  - nvm install --latest-npm
 
 env: PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
 
 jobs:
   include:
     - name: Lint
-      before_install:
-        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run lint
 
     - name: Build artifacts
-      before_install:
-        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run check-local-changes
 
     - name: License compatibility
-      before_install:
-        - nvm install --latest-npm
       install:
         - npm ci
       script:
         - npm run check-licenses
 
     - name: JavaScript unit tests
-      before_install:
-        - nvm install --latest-npm
       install:
         - npm ci
       script:
@@ -92,7 +84,8 @@ jobs:
     - name: E2E tests (Admin with plugins) (1/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
       install:
-        - ./bin/setup-local-env.sh
+        - npm ci
+        - ./bin/setup-travis-e2e-tests.sh
       script:
         - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
@@ -101,7 +94,8 @@ jobs:
     - name: E2E tests (Admin with plugins) (2/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
       install:
-        - ./bin/setup-local-env.sh
+        - npm ci
+        - ./bin/setup-travis-e2e-tests.sh
       script:
         - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
@@ -110,7 +104,8 @@ jobs:
     - name: E2E tests (Admin with plugins) (3/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
       install:
-        - ./bin/setup-local-env.sh
+        - npm ci
+        - ./bin/setup-travis-e2e-tests.sh
       script:
         - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
@@ -119,7 +114,8 @@ jobs:
     - name: E2E tests (Admin with plugins) (4/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false POPULAR_PLUGINS=true PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
       install:
-        - ./bin/setup-local-env.sh
+        - npm ci
+        - ./bin/setup-travis-e2e-tests.sh
       script:
         - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
@@ -128,7 +124,8 @@ jobs:
     - name: E2E tests (Author without plugins) (1/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
       install:
-        - ./bin/setup-local-env.sh
+        - npm ci
+        - ./bin/setup-travis-e2e-tests.sh
       script:
         - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
@@ -137,7 +134,8 @@ jobs:
     - name: E2E tests (Author without plugins) (2/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
       install:
-        - ./bin/setup-local-env.sh
+        - npm ci
+        - ./bin/setup-travis-e2e-tests.sh
       script:
         - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
@@ -146,7 +144,8 @@ jobs:
     - name: E2E tests (Author without plugins) (3/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
       install:
-        - ./bin/setup-local-env.sh
+        - npm ci
+        - ./bin/setup-travis-e2e-tests.sh
       script:
         - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests
@@ -155,7 +154,8 @@ jobs:
     - name: E2E tests (Author without plugins) (4/4)
       env: WP_VERSION=latest SCRIPT_DEBUG=false E2E_ROLE=author PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=
       install:
-        - ./bin/setup-local-env.sh
+        - npm ci
+        - ./bin/setup-travis-e2e-tests.sh
       script:
         - npm run build
         - $( npm bin )/wp-scripts test-e2e --config=./packages/e2e-tests/jest.config.js --listTests > ~/.jest-e2e-tests