--- a/bin/setup-local-env.sh
+++ b/bin/setup-local-env.sh
@@ -18,6 +18,19 @@ cd "$(dirname "$0")/.."
 # Check Docker is installed and running
 . "$(dirname "$0")/install-docker.sh"
 
+# Set up WordPress Development site.
+# Note: we don't bother installing the test site right now, because that's
+# done on every time `npm run test-e2e` is run.
+. "$(dirname "$0")/install-wordpress.sh"
+
+# Install the PHPUnit test scaffolding.
+echo -e $(status_message "Installing PHPUnit test scaffolding...")
+docker-compose $DOCKER_COMPOSE_FILE_OPTIONS run --rm wordpress_phpunit /app/bin/install-wp-tests.sh wordpress_test root example mysql $WP_VERSION false > /dev/null
+
+# Install Composer. This is only used to run WordPress Coding Standards checks.
+echo -e $(status_message "Installing and updating Composer modules...")
+docker-compose $DOCKER_COMPOSE_FILE_OPTIONS run --rm composer install
+
 ! read -d '' GUTENBERG <<"EOT"
 ,⁻⁻⁻·       .                 |
 |  ،⁓’.   . |---  ,---. ,---. |---. ,---. ,---. ,---.