--- a/bin/setup-travis-e2e-tests.sh
+++ b/bin/setup-travis-e2e-tests.sh
@@ -1,0 +1,26 @@
+#!/bin/bash
+
+# Exit if any command fails
+set -e
+
+# Set up environment variables
+. "$(dirname "$0")/bootstrap-env.sh"
+
+# Include useful functions
+. "$(dirname "$0")/includes.sh"
+
+# Change to the expected directory
+cd "$(dirname "$0")/.."
+
+# Download image updates.
+echo -e $(status_message "Downloading Docker image updates...")
+docker-compose $DOCKER_COMPOSE_FILE_OPTIONS pull mysql wordpress_e2e_tests cli_e2e_tests
+
+# Launch the containers.
+echo -e $(status_message "Starting Docker containers...")
+docker-compose $DOCKER_COMPOSE_FILE_OPTIONS up -d --remove-orphans mysql wordpress_e2e_tests cli_e2e_tests >/dev/null
+
+# Set up WordPress Development site.
+# Note: we don't bother installing the test site right now, because that's
+# done on every time `npm run test-e2e` is run.
+. "$(dirname "$0")/install-wordpress.sh" --e2e_tests