--- a/docker-compose-localdev.yml
+++ b/docker-compose-localdev.yml
@@ -2,7 +2,7 @@ version: '3.1'
 
 services:
 
-  wordpress:
+  wordpress_dev:
     volumes:
       - ./wordpress:/var/www/html