--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -2,20 +2,21 @@ version: '3.1'
 
 services:
 
-  wordpress:
+  wordpress_dev:
     image: wordpress
     restart: always
     ports:
       - 8888:80
     environment:
       WORDPRESS_DB_HOST: mysql
+      WORDPRESS_DB_NAME: wordpress_dev
       WORDPRESS_DB_PASSWORD: example
       ABSPATH: /usr/src/wordpress/
       WORDPRESS_DEBUG: 1
       WORDPRESS_CONFIG_EXTRA: |
         define( 'SCRIPT_DEBUG', true );
     volumes:
-      - wordpress_data:/var/www/html
+      - wordpress_dev_data:/var/www/html
       - .:/var/www/html/wp-content/plugins/gutenberg
       - ./packages/e2e-tests/plugins:/var/www/html/wp-content/plugins/gutenberg-test-plugins
       - ./packages/e2e-tests/mu-plugins:/var/www/html/wp-content/mu-plugins
@@ -27,11 +28,11 @@ services:
     restart: always
     user: xfs
     volumes:
-      - wordpress_data:/var/www/html
+      - wordpress_dev_data:/var/www/html
       - .:/var/www/html/wp-content/plugins/gutenberg
     depends_on:
       - mysql
-      - wordpress
+      - wordpress_dev
 
   mysql:
     image: mysql:5.7
@@ -91,5 +92,5 @@ services:
 
 volumes:
   testsuite:
-  wordpress_data:
+  wordpress_dev_data:
   wordpress_e2e_tests_data: