--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -2,21 +2,20 @@ version: '3.1'
 
 services:
 
-  wordpress_dev:
+  wordpress:
     image: wordpress
     restart: always
     ports:
       - 8888:80
     environment:
       WORDPRESS_DB_HOST: mysql
-      WORDPRESS_DB_NAME: wordpress_dev
       WORDPRESS_DB_PASSWORD: example
       ABSPATH: /usr/src/wordpress/
       WORDPRESS_DEBUG: 1
       WORDPRESS_CONFIG_EXTRA: |
         define( 'SCRIPT_DEBUG', true );
     volumes:
-      - wordpress_dev_data:/var/www/html
+      - wordpress_data:/var/www/html
       - .:/var/www/html/wp-content/plugins/gutenberg
       - ./packages/e2e-tests/plugins:/var/www/html/wp-content/plugins/gutenberg-test-plugins
       - ./packages/e2e-tests/mu-plugins:/var/www/html/wp-content/mu-plugins
@@ -28,11 +27,11 @@ services:
     restart: always
     user: xfs
     volumes:
-      - wordpress_dev_data:/var/www/html
+      - wordpress_data:/var/www/html
       - .:/var/www/html/wp-content/plugins/gutenberg
     depends_on:
       - mysql
-      - wordpress_dev
+      - wordpress
 
   mysql:
     image: mysql:5.7
@@ -92,5 +91,5 @@ services:
 
 volumes:
   testsuite:
-  wordpress_dev_data:
+  wordpress_data:
   wordpress_e2e_tests_data: