… bin/reset-local-e2e-tests.sh