## Description
This PR adds some improvements to the performance of Travis jobs as part of #15159.

The main difference in this proposal is that we no longer pull, create and start containers that are never used on Travis. This applies for jobs which run e2e tests and PHP unit tests.

| Job | [Before](https://travis-ci.com/WordPress/gutenberg/builds/112496088) | [After](https://travis-ci.com/WordPress/gutenberg/builds/112573336) |
|---|---|---|
| Lint | 1 min 55 sec | 2 min 3 sec |
| Build artifacts | 1 min 27 sec |  1 min 25 sec |
| License compatibility | 1 min 20 sec | 1 min 18 sec |
| JavaScript unit tests | 3 min 50 sec | 4 min 35 sec |
| PHP unit tests (Docker) | 5 min 59 sec | **3 min 42 sec** |
| E2E tests (Admin with plugins) (1/4) | 11 min 14 sec | **8 min 29 sec** |
| E2E tests (Admin with plugins) (2/4) | 10 min 48 sec | **7 min 21 sec** |
| E2E tests (Admin with plugins) (3/4) | 12 min 56 sec | **9 min 36 sec** |
| E2E tests (Admin with plugins) (4/4) | 10 min 46 sec |  **9 min 38 sec** |
| E2E tests (Author without plugins) (1/4 | 13 min 41 sec |  **12 min 26 sec** |
| E2E tests (Author without plugins) (2/4) | 8 min 39 sec |  **7 min 58 sec** |
| E2E tests (Author without plugins) (3/4) | 10 min 22 sec |  **10 min 4 sec** |
| E2E tests (Author without plugins) (4/4) | 9 min 20 sec | **8 min 34 sec** |
| PHP unit tests (PHP 5.2) | 3 min 22 sec |  3 min 15 sec |

It's very easy to notices that there is a huge gain on some of the jobs where they finish even 2-3 minutes faster!!!

Aside: I also tried to remove `npm ci` and `npm run build` from the jobs which run PHP unit tests but it turned out that it's still necessary for some tests to pass. See https://github.com/WordPress/gutenberg/issues/15159#issuecomment-494301504 for more details.

## How has this been tested?
Travis should pass all checks.